requirements.txt
numpy
python-dotenv
pandas
openpyxl
langdetect
streamlit
nltk
ensure
langchain_openai